# backend_monbrobroli-backend/storage/api-docs/api-docs.json
{
    "openapi": "3.0.0",
    "info": {
        "title": "L5 OpenApi",
        "description": "L5 Swagger OpenApi description",
        "contact": {
            "email": "darius@matulionis.lt"
        },
        "license": {
            "name": "Apache 2.0",
            "url": "https://www.apache.org/licenses/LICENSE-2.0.html"
        },
        "version": "1.0.0",
        "x-logo": {
            "url": "https://via.placeholder.com/190x90.png?text=L5-Swagger"
        }
    },
    "paths": {
...
        "api/admin/update_competence/{id}": {
            "put": {
                "tags": [
                    "Competences."
                ],
                "summary": "Update la compétence.",
                "operationId": "Update la compétence",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "description": "Id de la compétence.",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "requestBody": {
                    "description": "Modifier la compétence",
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "properties": {
                                    "competence": {
                                        "type": "string",
                                        "example": "Etudiant"
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "400": {
                        "description": "Invalid ID supplied"
                    },
                    "200": {
                        "description": "successful operation",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "status": {
                                            "type": "boolean",
                                            "example": true
                                        },
                                        "data": {
                                            "type": "object"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "404": {
                        "description": "Not found"
                    },
                    "405": {
                        "description": "Mauvaise méthode"
                    }
                }
            }
        },
        "api/admin/delete_categorie/{id}": {
            "delete": {
                "tags": [
                    "Categories."
                ],
                "summary": "Supprimer une catégorie",
                "description": "Supprimer une catégorie",
                "operationId": "delete_categorie",
                "parameters": [
                    {
                        "name": "Id",
                        "in": "path",
                        "description": "Id de la catégorie.",
                        "required": true,
                        "schema": {
                            "type": "string"
                        }
                    }
                ],
                "responses": {
                    "400": {
                        "description": "Invalid schedule supplied"
                    },
                    "200": {
                        "description": "successful operation",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "status": {
                                            "type": "boolean",
                                            "example": true
                                        },
                                        "message": {
                                            "type": "string",
                                            "example": "Catégorie supprimée avec succes"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "404": {
                        "description": "route not found"
                    }
                }
            }
        },
        "api/GetAllCompetences": {
            "get": {
                "tags": [
                    "Etudiants."
                ],
                "summary": "Voir toutes les compétences créées.",
                "description": "Retourne toutes les compétences créées",
                "operationId": "Voir toutes les compétences créées.",
                "responses": {
                    "200": {
                        "description": "successful operation",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "status": {
                                            "type": "boolean",
                                            "example": true
                                        },
                                        "data": {
                                            "type": "array",
                                            "items": {
                                                "type": "object"
                                            }
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "400": {
                        "description": "Bad request"
                    }
                }
            }
        },
...
        "api/toogleFavoris": {
            "post": {
                "tags": [
                    "Favoris-Student."
                ],
                "summary": "ajouter et rétirer les offres dans les favoris",
                "description": "ajouter et rétirer les offres dans les favoris",
                "operationId": "bb0237cae84e4c2cab65c94f44a64cf2",
                "requestBody": {
                    "description": "Evaluation",
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "properties": {
                                    "offre_id": {
                                        "type": "integer",
                                        "example": "5"
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "successful operation",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "status": {
                                            "type": "boolean",
                                            "example": true
                                        },
                                        "message": {
                                            "type": "string",
                                            "example": "...."
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    }
                }
            }
        },
        "api/listFavoris": {
            "get": {
                "tags": [
                    "Favoris-Student."
                ],
                "summary": "Voir la liste de mes favoris",
                "description": "Retourne la liste des favoris.",
                "operationId": "Voir la liste de mes favoris.",
                "responses": {
                    "200": {
                        "description": "successful operation",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "status": {
                                            "type": "boolean",
                                            "example": true
                                        },
                                        "data": {
                                            "type": "array",
                                            "items": {
                                                "type": "object"
                                            }
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "400": {
                        "description": "Bad request"
                    }
                }
            }
        },
...
    },
...
}